fix warnings

+++ include/ec_mailbox.h
@@ -22,7 +22,7 @@ int ec_mailbox_receive(ec_master_t *master,
                        uint16_t slave_index,
                        ec_datagram_t *datagram,
                        uint8_t *type,
-                       uint32_t *size,
+                       uint16_t *size,
                        uint64_t timeout_ns);
 
 #endif
